fix different device error on sdxl models #337
Fixes an error with sdxl models when merging using trainDifference. When
there is insufficient VRAM, each tensor may be located on different devices.
These must be swapped to the correct device when operations are performed on
the tensors. traindiff now ensures that tensors are always loaded onto the
correct device to prevent the error from occurring.